Separate out test base file
